fix: repair seach function by move router.
