+++ routes/index.js
@@ -12,5 +12,10 @@ const restaurants = require('./modules/restaurants')
 // 將網址結構符合 /restaurants 字串開頭的 request 導向 restaurants 模組 
 router.use('/restaurants', restaurants)
 
+const search = require('./modules/search')
+// 將網址結構符合 /search 字串開頭導向 search模組 
+router.use('/restaurants', search)
+
+
 // 匯出路由器
 module.exports = router
